build(image): added manual keys for external repos

+++ scripts/deps/repo.sh
@@ -40,9 +40,16 @@ disable_repo() {
 enable_repo_extern() {
     local repo_file_name=$1
     local repo_name=$2
+    local repo_key
+
+    repo_key="${repo_file_name##*/}"
+    repo_key="RPM-GPG-KEY-${repo_key%.repo}"
 
     local repo_file_source="${REPOS_DIR}/${repo_file_name}"
     local repo_file_destination="/etc/yum.repos.d/${repo_file_name}"
+
+    local repo_key_file_source="${REPO_KEYS_DIR}/${repo_key}"
+    local repo_key_file_destination="/etc/pki/rpm-gpg/${repo_key}"
 
     echo_color "Installing external repo '$repo_name' in $repo_file_source"
 
@@ -50,8 +57,13 @@ enable_repo_extern() {
         echo_error "Repo source file for '$repo_name' not found: $repo_file_source"
         return 1
     fi
+    if [ ! -f "$repo_key_file_source" ]; then
+        echo_error "Repo key file for '$repo_name' not found: $repo_key_file_source"
+        return 1
+    fi
 
     cp "$repo_file_source" "$repo_file_destination"
+    cp "$repo_key_file_source" "$repo_key_file_destination"
 
     enable_repo "$repo_file_name" "$repo_name"
 }
@@ -60,9 +72,14 @@ disable_repo_extern() {
     local repo_file_name=$1
     local repo_name=$2
 
+    repo_key="${repo_file_name##*/}"
+    repo_key="RPM-GPG-KEY-${repo_key%.repo}"
+
     local repo_file_destination="/etc/yum.repos.d/${repo_file_name}"
+    local repo_key_file_destination="/etc/pki/rpm-gpg/${repo_key}"
 
     echo_color "Removing external repo '$repo_name' in $repo_file_destination"
 
     rm "$repo_file_destination"
+    rm "$repo_key_file_destination"
 }
